fix(backup-script): ensure backup file is only removed after successful upload
- Add condition to check `UPLOAD_TO_S3_FINISHED_SUCCESSFULLY` before removing the backup file
- Prevents premature removal of the backup file if the upload has not completed successfully

+++ .docker/aws-cli/etc/backup.sh
@@ -127,7 +127,7 @@ if [ "${REMOVE_DATABASE_DUMP_FILES}" = 1 ]; then
   fi
 fi
 
-if [ "${REMOVE_BACKUP_FILE}" = 1 ]; then
+if [ "${REMOVE_BACKUP_FILE}" = 1 ] && [ -n "${UPLOAD_TO_S3_FINISHED_SUCCESSFULLY}" ]; then
   if [ -f "${BACKUP_FILE_PATH}" ]; then
     echo " 🗑️ Removing ${BACKUP_FILE_PATH}"
     rm -f "${BACKUP_FILE_PATH}"
